requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.1
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.10.0
django-timezone-field==5.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
kombu==5.3.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
wfastcgi==3.0.0
zope.event==5.0
zope.interface==6.0
